+++ moca_pop/aux_scripts/pipeline_runner/config_builder/nushu_pipeline_series.py
@@ -32,7 +32,7 @@ import os
 
 from moca_pop.config import directory
 from moca_pop.utils import json_utils
-import moca_pop.scripts.pipeline_runner.pipeline as pipeline
+import moca_pop.aux_scripts.pipeline_runner.pipeline as pipeline
 
 
 rl_pipeline = pipeline.Pipeline(
@@ -96,7 +96,7 @@ nushu_pipeline_series = pipeline.PipelineSeries(
 
 
 model_json = nushu_pipeline_series.model_dump_json(indent=4)
-output_dir = os.path.join(directory.SCRIPTS_DIR, "pipeline_runner", "config")
+output_dir = os.path.join(directory.AUX_DIR, "pipeline_runner", "config")
 output_fn = os.path.join(output_dir, "nushu_pipeline_series.json")
 
 with open(output_fn, "w") as f: